made tripController
