+++ src/main.js
@@ -2,13 +2,11 @@ import {tripInfo as tripInfoLayout} from './components/tripInfo.js';
 import {menu as menuLayout} from './components/menu.js';
 import {filtersLayout} from './components/filters.js';
 
-import Event from './components/eventItem';
-import EventEdit from './components/eventEdit';
-
 import {sort as sortLayout} from './components/sort.js';
 import {getPoint, getFilter, getTripInfo, getDays} from './data.js';
 import {renderTripDay} from './renderTripDay.js';
-import {render, Position, unRender} from './utils';
+
+import TripController from './components/tripController';
 
 const mockPoints = new Array(3).fill(``).map(() => getPoint(Math.floor(Math.random() * 100), {type: `activity`, number: 2}));
 
@@ -45,59 +43,5 @@ tripEvents.appendChild(tripDays);
 
 const eventsContainer = document.querySelector(`.trip-events__list`);
 
-const renderPoint = (mockPoint) => {
-  const event = new Event(mockPoint);
-  const eventEdit = new EventEdit(mockPoint);
-
-  const hideOnEsc = (evt) => {
-    if (evt.key === `Escape` || evt.key === `Esc`) {
-      eventsContainer.replaceChild(event.getElement(), eventEdit.getElement());
-    }
-  }; 
-
-  event.getElement()
-    .querySelector(`.event__rollup-btn`)
-    .addEventListener(`click`, () => {
-      eventsContainer.replaceChild(eventEdit.getElement(), event.getElement());
-      document.addEventListener(`keydown`, hideOnEsc);
-    });
-
-  eventEdit.getElement()
-    .querySelector(`.event__rollup-btn`)
-    .addEventListener(`click`, () => {
-      eventsContainer.replaceChild(event.getElement(), eventEdit.getElement());
-      document.removeEventListener(`keydown`, hideOnEsc);
-    });
-  eventEdit.getElement()
-    .querySelector(`.event__save-btn`)
-    .addEventListener(`click`, () => {
-      eventsContainer.replaceChild(event.getElement(), eventEdit.getElement());
-      document.removeEventListener(`keydown`, hideOnEsc);
-    });
-
-  eventEdit.getElement()
-    .querySelector(`form`)
-    .addEventListener(`submit`, () => {
-      eventsContainer.replaceChild(event.getElement(), eventEdit.getElement());
-      document.removeEventListener(`keydown`, hideOnEsc);
-    });
-
-  eventEdit.getElement()
-    .querySelector(`.event__reset-btn`)
-    .addEventListener(`click`, () => {
-      unRender(eventEdit.getElement());
-      eventEdit.removeElement();
-      document.removeEventListener(`keydown`, hideOnEsc);
-    });
-
-  render(eventsContainer, event.getElement(), Position.BEFOREEND);
-};
-
-const futurePoints = mockPoints.filter((day) => day.timeStart > Date.now());
-console.log(futurePoints);
-
-if (futurePoints.length === 0) {
-  tripEvents.innerHTML = `<h2 class="visually-hidden">Trip events</h2><p class="trip-events__msg">Click New Event to create your first point</p>`;
-} else {
-  mockPoints.map((point) => renderPoint(point));
-}
+const tripController = new TripController(eventsContainer, mockPoints);
+tripController.init();
